refactor: introduce `inventory` module
I was finding the main `lib` module difficult to navigate, so I decided to move inventory-related
code into its own module, and introduced a new `DeploymentInventoryService` struct for encapsulating
the generation of inventory.

I also made changes and additions in the `ansible` module, introducing concepts from our own domain
rather than making it completely generic. We now pass enums for playbooks and inventory values, so
we don't have to assemble paths all over the place. I think this makes the code more compact,
readable and maintainable.
